Research Information is added
