+++ Platon/backend/app/profile_management/views.py
@@ -1,8 +1,6 @@
 from flask import make_response,jsonify,request
-from flask_restplus import Resource
+from flask_restplus import Resource,Namespace
 from enum import Enum
-from functools import wraps
-from scholarly import scholarly
 import requests
 import json
 import atexit
@@ -10,6 +8,7 @@ from apscheduler.schedulers.background import BackgroundScheduler
 
 from app.auth_system.views import login_required
 from app.auth_system.models import User
+from app.follow_system.views import follow_required
 from app.follow_system.models import Follow
 from app.profile_management.forms import ResearchInfoGetForm,research_info_get_parser, ResearchInfoPostForm,research_info_post_parser
 from app.profile_management.forms import ResearchInfoUpdateForm, research_info_update_parser,ResearchInfoDeleteFrom,research_info_delete_parser
@@ -20,44 +19,11 @@ class ResearchType(Enum):
     HAND_WRITTEN = 0
     FETCHED = 1
 
-def follow_required(param_loc,requested_user_id_key):
-    """
-        param_loc: it only can be "args" or "form" which specifies the location of the following id parameter in a request
-        requested_user_id_key: key of the parameter that represents the requested user id in the request
-    """
-    def follow_required_inner(func):
-        """
-            Checks the privacy of the user before giving the user data
-        """
-        @wraps(func)
-        def follow_check(*args,**kwargs):
-            if param_loc == 'args':
-                request_dict = request.args
-            elif param_loc == 'form':
-                request_dict = request.form
-            if requested_user_id_key not in request_dict:
-                return make_response(jsonify({'error':'Wrong input format'}),400)
-            else:
-                requested_user_id = request_dict.get(requested_user_id_key)
-                user_id = args[0]
-            try:
-                requested_user = User.query.filter(User.id == requested_user_id and User.is_valid == True).first()
-            except:
-                return make_response(jsonify({'error' : 'Database Connection Problem'}),500)
-            if requested_user is None:
-                return make_response(jsonify({'error': 'Requested user can not be found'}),404)
-            if not requested_user.is_private:
-                try:
-                    follow = Follow.query.filter(Follow.follower_id == user_id and Follow.following_id == requested_user.id).first()
-                except:
-                    return make_response(jsonify({'error' : 'Database Connection Problem'}),500)
-                if follow is None:
-                    return make_response(jsonify({'error' : 'The Account that you try to reach is private'}),403)
-            return func(*args,**kwargs)
-        return follow_check
-    return follow_required_inner
+profile_management_ns = Namespace("Profile Management",
+                                  description="Profile Management Endpoints",
+                                  path = "/profile")
 
-    
+@profile_management_ns.route("/research_information")
 class ResearchInformationAPI(Resource):
     """
         Research Information Functionality is implemented in this class
@@ -84,6 +50,9 @@ class ResearchInformationAPI(Resource):
     @api.expect(research_info_post_parser)
     @login_required
     def post(user_id,self):
+        """
+            Adds new Research Information
+        """
         form = ResearchInfoPostForm(request.form)
         if form.validate():
             new_research_info = ResearchInformation(user_id,form.research_title.data,form.description.data,form.year.data,ResearchType.HAND_WRITTEN)
@@ -99,7 +68,10 @@ class ResearchInformationAPI(Resource):
     @api.doc(responses={201:'Successfully Updated',400:'Wrong Input Format',401:'Authantication Problem',500:'Database Connection Problem'})
     @api.expect(research_info_update_parser)
     @login_required
-    def update(user_id,self):
+    def put(user_id,self):
+        """
+            Updates a Research Information
+        """
         form = ResearchInfoUpdateForm(request.form)
         if form.validate():
             try:
@@ -121,6 +93,9 @@ class ResearchInformationAPI(Resource):
     @api.expect(research_info_delete_parser)
     @login_required
     def delete(user_id,self):
+        """
+            Deletes Research Information
+        """
         form = ResearchInfoDeleteFrom(request.args)
         if form.validate():
             try:
@@ -135,12 +110,13 @@ class ResearchInfoFetch():
     
     @staticmethod
     def fetch_google_scholar_info(username):
+        api_url = "http://cse.bth.se/~fer/googlescholar-api/googlescholar.php?user={}".format(username)
         try:
-            search_query = scholarly.search_author(username)
-            author = next(search_query).fill()
+            response = requests.get(api_url)
+            response = json.loads(response.text)
+            return [{'title':research['title'],'description':'','year':research['year']}for research in response['publications']]
         except:
             return []
-        return [{'title':i.bib['title'],'description':'','year':int(i.bib['year'])}for i in author.publications]
     
     @staticmethod
     def fetch_research_gate_info(username):
@@ -155,6 +131,7 @@ class ResearchInfoFetch():
     
     @staticmethod
     def update_research_info():
+        print("deneme")
         try:
             all_users = User.query.all()
         except:
@@ -176,10 +153,11 @@ class ResearchInfoFetch():
 
 def schedule_regularly():
     scheduler = BackgroundScheduler()
-    scheduler.add_job(func=ResearchInfoFetch.update_research_info, trigger="interval",hour=1)
+    scheduler.add_job(func=ResearchInfoFetch.update_research_info, trigger="interval",seconds=60*60)
     scheduler.start()
     # Shut down the scheduler when exiting the app
     atexit.register(lambda: scheduler.shutdown())
 
+
 def register_resources(api):
-    api.add_resource(ResearchInformationAPI,"/profile/research_information")
+    api.add_namespace(profile_management_ns)
